+++ src/frontend_analysis/IR_analysis/InterfaceInfer.hpp
@@ -95,27 +95,31 @@ class InterfaceInfer : public ApplicationFrontendFlowStep
 
    void create_resource_Read_simple(const std::set<std::string>& operations, const std::string& arg_name,
                                     const std::string& interfaceType, unsigned long long input_bw, bool IO_port,
-                                    unsigned n_resources, unsigned long long rwBWsize, unsigned int top_id) const;
+                                    unsigned long long n_resources, unsigned long long rwBWsize,
+                                    unsigned int top_id) const;
 
    void create_resource_Write_simple(const std::set<std::string>& operations, const std::string& arg_name,
                                      const std::string& interfaceType, unsigned long long input_bw, bool IO_port,
-                                     bool isDiffSize, unsigned n_resources, bool is_real, unsigned long long rwBWsize,
-                                     unsigned int top_id) const;
+                                     bool isDiffSize, unsigned long long n_resources, bool is_real,
+                                     unsigned long long rwBWsize, unsigned int top_id) const;
 
    void create_resource_array(const std::set<std::string>& operationsR, const std::set<std::string>& operationsW,
                               const std::string& bundle_name, const std::string& interfaceType,
                               unsigned long long input_bw, unsigned int arraySize, unsigned n_resources,
-                              unsigned long long alignment, bool is_real, unsigned long long rwBWsize, unsigned int top_id) const;
+                              unsigned long long alignment, bool is_real, unsigned long long rwBWsize,
+                              unsigned int top_id) const;
 
    void create_resource_m_axi(const std::set<std::string>& operationsR, const std::set<std::string>& operationsW,
                               const std::string& arg_name, const std::string& bundle_name,
-                              const std::string& interfaceType, unsigned long long input_bw, unsigned n_resources,
-                              m_axi_type mat, unsigned long long rwBWsize, unsigned int top_id) const;
+                              const std::string& interfaceType, unsigned long long input_bw,
+                              unsigned long long n_resources, m_axi_type mat, unsigned long long rwBWsize,
+                              unsigned int top_id) const;
 
    void create_resource(const std::set<std::string>& operationsR, const std::set<std::string>& operationsW,
                         const std::string& arg_name, const std::string& interfaceType, unsigned long long input_bw,
-                        bool isDiffSize, const std::string& fname, unsigned long long n_resources, unsigned long long alignment,
-                        bool isReal, unsigned long long rwBWsize, unsigned int top_id) const;
+                        bool isDiffSize, const std::string& fname, unsigned long long n_resources,
+                        unsigned long long alignment, bool isReal, unsigned long long rwBWsize,
+                        unsigned int top_id) const;
 
    void ComputeResourcesAlignment(unsigned long long& n_resources, unsigned long long& alignment,
                                   unsigned long long input_bw, bool is_acType, bool is_signed, bool is_fixed);
